ml/requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0
scikit-learn>=0.24.0
tensorflow>=2.10.0
keras>=2.10.0
opencv-python>=4.5.0
requests>=2.27.0
paho-mqtt>=1.6.0
python-dotenv>=0.19.0
flask>=2.0.0
flask-cors>=3.0.0
joblib>=1.1.0
matplotlib>=3.4.0
seaborn>=0.11.0
